Moved argument parsing for Combobulator
