+++ src/combobulator.py
@@ -6,7 +6,6 @@
     Returns:
         int: Exit code
 """
-import argparse
 import os
 import csv
 import sys
@@ -20,6 +19,7 @@ from registry import maven
 from registry import pypi
 from analysis import heuristics as heur
 from constants import ExitCodes, PackageManagers, Constants  # Import Constants including LOG_FORMAT
+from args import parse_args
 
 SUPPORTED_PACKAGES = Constants.SUPPORTED_PACKAGES
 
@@ -28,75 +28,6 @@ def init_args():
     # WARNING: don't populate this instance with a hard-coded value
     # it is merely for initializing a string variable.
     GITHUB_TOKEN=""
-
-def parse_args():
-    """Parses the arguments passed to the program."""
-    parser = argparse.ArgumentParser(
-        prog="combobulator.py",
-        description="Dependency Combobulator - Dependency Confusion Checker",
-        epilog='Apiiro <Heart> Community',
-        add_help=True)
-    parser.add_argument("-t", "--type",
-                        dest="package_type",
-                        help="Package Manager Type, i.e: npm, PyPI, maven",
-                        action="store",type=str, choices=SUPPORTED_PACKAGES,
-                        required=True )
-    # https://docs.python.org/3/library/argparse.html#mutual-exclusion
-    # input_group as a mutually exclusive arg group: 
-    input_group = parser.add_mutually_exclusive_group(required=True)
-    input_group.add_argument("-l", "--load_list",
-                        dest="LIST_FROM_FILE",
-                        help="Load list of dependencies from a file",
-                        action="append",type=str,
-                        default=[] )
-    input_group.add_argument("-d", "--directory",
-                    dest="FROM_SRC",
-                    help="Extract dependencies from local source repository",
-                    action="append",
-                    type=str)
-    input_group.add_argument("-p", "--package",
-                            dest="SINGLE",
-                            help="Name a single package.",
-                            action="append",type=str )
-    output_group = parser.add_mutually_exclusive_group(required=False)
-    output_group.add_argument("-c", "--csv",
-        dest="CSV",
-        help="Export packages properties onto CSV file",
-                    action="store", type=str)
-    # support variables
-    parser.add_argument("-gh", "--github",
-                    dest="GITHUB_TOKEN",
-                    help="GitHub Access Token (Overrides .env file setting)",
-                    action="store", type=str )
-    parser.add_argument("-a", "--analysis",
-        dest="LEVEL",
-        help="Required analysis level - compare (comp), heuristics (heur) (default: compare)",
-                    action="store", default="compare", type=str,
-                    choices = Constants.LEVELS)
-    # Added new arguments for logging
-    parser.add_argument("--loglevel",
-                        dest="LOG_LEVEL",
-                        help="Set the logging level",
-                        action="store",
-                        type=str,
-                        choices=['DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL'],
-                        default='INFO')
-    parser.add_argument("--logfile",
-                        dest="LOG_FILE",
-                        help="Log output file",
-                        action="store",
-                        type=str)
-    parser.add_argument("-r", "--recursive",
-                        dest="RECURSIVE",
-                        help="Recursively scan directories when scanning from source.",
-                        action="store_true")
-    # Add new argument for controlling exit on warnings
-    parser.add_argument("--error-on-warnings",
-                        dest="ERROR_ON_WARNINGS",
-                        help="Exit with a non-zero status code if warnings are present.",
-                        action="store_true")
-    return parser.parse_args()
-
 
 def load_env():
     """
@@ -288,4 +219,3 @@ def main():
 
 if __name__ == "__main__":
     main()
-
